Moved formatting code out of template and added progress to DownloadMonitor.vue
