+++ src/pages/DownloadMonitor.vue
@@ -28,6 +28,7 @@
                             <div class="border-at-bottom pad pad--sides">
                                 <div class="card is-shadowless">
                                     <p><strong>{{ downloadObject.initialMods.join(", ") }}</strong></p>
+                                    <p>{{convertSizeToReadable(downloadObject.currentProgress, false)}}/{{convertSizeToReadable(downloadObject.currentModSize, true)}}</p>
                                     <p v-if="downloadObject.progress < 100">Downloading: {{ downloadObject.modName }}</p>
                                     <p>{{Math.min(Math.floor(downloadObject.progress), 100)}}% complete</p>
                                     <Progress v-if="!downloadObject.failed"
@@ -57,6 +58,7 @@ import { Hero } from '../components/all';
 import NavigationMenuProvider from '../providers/components/loaders/NavigationMenuProvider';
 import DownloadModModal from '../components/views/DownloadModModal.vue';
 import Progress from '../components/Progress.vue';
+import FormatUtils from '../utils/FormatUtils';
 import Timeout = NodeJS.Timeout;
 
 @Component({
@@ -79,6 +81,10 @@ export default class DownloadMonitor extends Vue {
 
     route(ref: string) {
         this.$router.replace(`/manager?view=${ref}`);
+    }
+
+    private convertSizeToReadable(size: number, withSizeIndicator: boolean) {
+        return FormatUtils.convertSizeToReadable(size, withSizeIndicator);
     }
 
     goto(ref: string) {
